\name{fixef}
\alias{fixef}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Method for extracting fixed effects}
\description{
 This is a generic function.}

\usage{
fixef(x, ...)}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{a fitted object from which fixed effects can be extracted.}
  \item{\dots}{may be required by some methods using this generic function.}
}

\value{
see the documentation for the specific class.}

\seealso{
 \code{\link{ranef}}}


\examples{
## See the specific class documentation.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{models}
