\name{vcov}
\alias{vcov}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Method for extracting the covariance matrix}
\description{
 This is a generic function.}

\usage{
vcov(object, ...)}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{a fitted object from which the covariance matrix can be extracted.}
  \item{\dots}{may be required by some methods using this generic function.}
}

\value{
see the documentation for the specific class.}


\examples{
## See the specific class documentation.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{models}
